feat(search-shared): add lexical-artifact and lexical-doc-id
Move lexical artifact/doc id helpers into search-shared; slim indexer
lexical-search modules to re-export or delegate.

Made-with: Cursor
